requirements.txt
python-docx == 0.8.10
google-trans-new == 1.1.9
